Int32 indexing in LinSolverCache, to match Jacobian functor

+++ src/LinearSolverCache/linear_solver_cache.jl
@@ -1,14 +1,14 @@
 abstract type LinearSolverCache end
 
-function symbolic_factor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int64})
+function symbolic_factor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int32})
     throw(AbstractMethodError(:symbolic_factor!))
 end
 
-function symbolic_refactor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int64})
+function symbolic_refactor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int32})
     throw(AbstractMethodError(:symbolic_refactor!))
 end
 
-function numeric_refactor!(cache::LinearSolverCache,  A::SparseMatrixCSC{Float64, Int64})
+function numeric_refactor!(cache::LinearSolverCache,  A::SparseMatrixCSC{Float64, Int32})
     throw(AbstractMethodError(:numeric_refactor!))
 end
 
@@ -16,13 +16,13 @@ function solve!(cache::LinearSolverCache, B::Vector{Float64})
     throw(AbstractMethodError(:solve!))
 end
 
-function full_refactor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int64})
+function full_refactor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int32})
     symbolic_refactor!(cache, A)
     numeric_refactor!(cache, A)
     return
 end
 
-function full_factor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int64})
+function full_factor!(cache::LinearSolverCache, A::SparseMatrixCSC{Float64, Int32})
     symbolic_factor!(cache, A)
     numeric_refactor!(cache, A)
     return
